app/requirements.txt
streamlit==1.31.0
pandas==2.3
pyspark==4.0
psycopg2-binary==2.9.11